+++ src/main/java/com/insane96mcp/shieldsplus/setup/Strings.java
@@ -12,12 +12,6 @@ public class Strings {
         public static final String NETHERITE_SHIELD = "netherite_shield";
     }
 
-    public static class Tags {
-        public static final String SHIELD_BASHING = ShieldsPlus.RESOURCE_PREFIX + "shield_bashing";
-        public static final String BASH_TIMER = ShieldsPlus.RESOURCE_PREFIX + "bash_timer";
-
-    }
-
     public static class Translatable {
         public static final String DAMAGE_BLOCKED = ShieldsPlus.RESOURCE_PREFIX + "damage_blocked";
         public static final String REINFORCED_BONUS = ShieldsPlus.RESOURCE_PREFIX + "reinforced_bonus";