+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldBashEnchantment.java
@@ -1,10 +1,9 @@
 package com.insane96mcp.shieldsplus.world.item.enchantment;
 
+import com.insane96mcp.shieldsplus.ShieldsPlus;
 import com.insane96mcp.shieldsplus.setup.SPEnchantments;
-import com.insane96mcp.shieldsplus.setup.Strings;
 import com.insane96mcp.shieldsplus.world.item.SPShieldItem;
 import net.minecraft.core.particles.ParticleTypes;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.util.Mth;
 import net.minecraft.world.damagesource.DamageSource;
@@ -19,6 +18,10 @@ import java.util.List;
 
 public class ShieldBashEnchantment extends Enchantment {
 
+	public static final String SHIELD_BASHING = ShieldsPlus.RESOURCE_PREFIX + "shield_bashing";
+	public static final String SHIELD_BASHING_LEVEL = ShieldsPlus.RESOURCE_PREFIX + "shield_bashing_level";
+	public static final String CHARGE_UP_TIMER = ShieldsPlus.RESOURCE_PREFIX + "charge_up_timer";
+
 	public ShieldBashEnchantment() {
 		super(Rarity.VERY_RARE, SPShieldItem.SHIELD, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
 	}
@@ -32,7 +35,7 @@ public class ShieldBashEnchantment extends Enchantment {
 	}
 
 	public int getMaxLevel() {
-		return 1;
+		return 3;
 	}
 
 	@Override
@@ -45,52 +48,95 @@ public class ShieldBashEnchantment extends Enchantment {
 		return !(enchantment instanceof ShieldReflectionEnchantment) && !(enchantment instanceof ShieldRecoilEnchantment) && super.checkCompatibility(enchantment);
 	}
 
+	public static double getForce(byte level) {
+		return 3d + (0.5d * (level - 1));
+	}
+
 	public static void onTick(Player player) {
-		CompoundTag tag = player.getPersistentData();
-		int shieldBash = player.getUseItem().getEnchantmentLevel(SPEnchantments.SHIELD_BASH.get());
+		byte shieldBash = (byte) player.getUseItem().getEnchantmentLevel(SPEnchantments.SHIELD_BASH.get());
 
-		if (tag.getByte(Strings.Tags.BASH_TIMER) < 0){
-			tag.putByte(Strings.Tags.BASH_TIMER, (byte) (tag.getByte(Strings.Tags.BASH_TIMER) + 1));
-		}
+		cooldownTickChargeUpTimer(player);
 
-		if (player.isBlocking() && tag.getByte(Strings.Tags.SHIELD_BASHING) <= 0 && shieldBash > 0) {
-			if (tag.getByte(Strings.Tags.BASH_TIMER) < 30) {
-				tag.putByte(Strings.Tags.BASH_TIMER, (byte) (tag.getByte(Strings.Tags.BASH_TIMER) + 1));
-				if (tag.getByte(Strings.Tags.BASH_TIMER) >= 30)
+		if (player.isBlocking() && getBashingTimer(player) <= 0 && shieldBash > 0) {
+			if (getChargeUpTimer(player) < 30) {
+				incrementChargeUpTimer(player);
+				if (getChargeUpTimer(player) >= 30)
 					player.playSound(SoundEvents.SHIELD_BLOCK, 1f, 1.65f);
 			}
 			else if (player.isCrouching() && player.isOnGround()) {
-				tag.putByte(Strings.Tags.SHIELD_BASHING, (byte) 12);
-				float f = -Mth.sin(player.getYRot() * ((float) Math.PI / 180F));
-				float f1 = Mth.cos(player.getYRot() * ((float) Math.PI / 180F));
+				setBashingTimer(player, (byte) 12);
+				setBashingLevel(player, shieldBash);
+				float x = -Mth.sin(player.getYRot() * ((float) Math.PI / 180F));
+				float z = Mth.cos(player.getYRot() * ((float) Math.PI / 180F));
 				player.playSound(SoundEvents.SHIELD_BLOCK, 1f, 1.3f);
-				player.setDeltaMovement(player.getDeltaMovement().add(f * 3d, 0.45d, f1 * 3d));
+				player.setDeltaMovement(player.getDeltaMovement().add(x * getForce(shieldBash), 0.45d + (shieldBash * 0.05d), z * getForce(shieldBash)));
 				for (int i = 0; i < 50; i++) {
 					player.level.addParticle(ParticleTypes.CLOUD, player.getX() + Mth.nextDouble(player.getRandom(), -0.5d, 0.5d), player.getY() + Mth.nextDouble(player.getRandom(), -0.5d, 0.5d) + 0.9d, player.getZ() + Mth.nextDouble(player.getRandom(), -0.5d, 0.5d), 0.1, 0, 0.1);
 				}
-				tag.putByte(Strings.Tags.BASH_TIMER, (byte) -30);
+				putChargeUpTimerOnCooldown(player);
 			}
 		}
-		else if (tag.getByte(Strings.Tags.BASH_TIMER) > 0) {
-			tag.putByte(Strings.Tags.BASH_TIMER, (byte) 0);
+		//If not blocking reset charge up cooldown
+		else if (getChargeUpTimer(player) > 0) {
+			setChargeUpTimer(player, (byte) 0);
 		}
 
-		if (tag.getByte(Strings.Tags.SHIELD_BASHING) > 0) {
-			tag.putByte(Strings.Tags.SHIELD_BASHING, (byte) (tag.getByte(Strings.Tags.SHIELD_BASHING) - 1));
+		if (getBashingTimer(player) > 0) {
+			setBashingTimer(player, (byte) (getBashingTimer(player) - 1));
 			if (player.isBlocking()) {
-				damageAndKnockback(player);
+				damageAndKnockback(player, getBashingLevel(player));
 			}
+			if (getBashingTimer(player) <= 0)
+				setBashingLevel(player, (byte) 0);
 		}
 	}
 
-	public static void damageAndKnockback(Player player) {
+	public static byte getChargeUpTimer(Player player) {
+		return player.getPersistentData().getByte(CHARGE_UP_TIMER);
+	}
+
+	public static void setChargeUpTimer(Player player, byte timer) {
+		player.getPersistentData().putByte(CHARGE_UP_TIMER, timer);
+	}
+
+	public static void cooldownTickChargeUpTimer(Player player) {
+		if (getChargeUpTimer(player) < 0)
+			incrementChargeUpTimer(player);
+	}
+
+	public static void incrementChargeUpTimer(Player player) {
+		setChargeUpTimer(player, (byte) (getChargeUpTimer(player) + 1));
+	}
+
+	public static void putChargeUpTimerOnCooldown(Player player) {
+		setChargeUpTimer(player, (byte) -30);
+	}
+
+	public static byte getBashingTimer(Player player) {
+		return player.getPersistentData().getByte(SHIELD_BASHING);
+	}
+
+	public static void setBashingTimer(Player player, byte timer) {
+		player.getPersistentData().putByte(SHIELD_BASHING, timer);
+	}
+
+	public static byte getBashingLevel(Player player) {
+		return player.getPersistentData().getByte(SHIELD_BASHING_LEVEL);
+	}
+
+	public static void setBashingLevel(Player player, byte level) {
+		player.getPersistentData().putByte(SHIELD_BASHING_LEVEL, level);
+	}
+
+	public static void damageAndKnockback(Player player, int level) {
 		AABB aabb = player.getBoundingBox().inflate(0.6d, 0.2d, 0.6d);
 		List<LivingEntity> entities = player.level.getEntitiesOfClass(LivingEntity.class, aabb, entity -> entity != player);
 		for (LivingEntity entity : entities) {
 			if (entity.isDeadOrDying() || entity.invulnerableTime > 10)
 				continue;
-			entity.knockback(1.5d, player.getX() - entity.getX(), player.getZ() - entity.getZ());
-			if (entity.hurt(DamageSource.playerAttack(player), 6)) {
+			entity.knockback(1d + (0.4d * level), player.getX() - entity.getX(), player.getZ() - entity.getZ());
+			//TODO Send knockback packet to player
+			if (entity.hurt(DamageSource.playerAttack(player), 3 + (3 * level))) {
 				player.playSound(SoundEvents.SHIELD_BLOCK, 1.0f, 0.5f);
 				ShieldAblazeEnchantment.apply(player, entity);
 			}